Breaking Changes to formatters
In order to support better metrics collection around dropped metrics, I
needed to have the format_stat/10 function return a value of
{ok, ResultsList, NumGood, NumBad}
The ResultsList can be [] if all results were bad.

This is mostly to support the RRD backend which can get errors from
cache lookups which I'd like to track.

Also, I modified the mondemand_backend_lwes to not deserialize and
reserialize if there is no extra context.  This was a regression as
the old server had the default for the extra context be undefined
while now it is [].