Fix bug when determining difference for mostly zero counters
When you have a counter that starts at 0 and rarely increases, when it
does increase the first time it would be silently dropped, because of
the clauses that checked if either was zero and just returned zero.  I
believe the code now works correctly even through a counter reset, and
added a test to show it (for once).