Fix to remove spikes on initial values after restart
When the server is restarted with the previous logic, you might get a
large value for a counter, then return that as the gauge value since the
values were defaulting to 0.  By changing the defaults to undefined and
have a clause to check if value2 is undefined, we ensure that we have to
have seen 2 values before we calculate the difference.  This removes
initial spikes on server restart and instead you just get 0 values for
that minute.